// src/ui/components/glyphs/typings/glyph-name.ts
export type GlyphName =
  | "add"
  | "apps"
  | "arrow-left"
  | "arrow-right"
  | "back-android"
  | "back-ios"
  | "business"
  | "calendar"
  | "car"
  | "checkmark"
  | "chemistry"
  | "clock-in-out"
  | "collapse"
  | "color-add"
  | "color-delete"
  | "color-edit"
  | "color"
  | "download"
  | "edit"
  | "ellipsis"
  | "email-send"
  | "email"
  | "expand"
  | "export"
  | "field"
  | "filter-clear"
  | "filter"
  | "flag"
  | "help"
  | "map"
  | "menu"
  | "message-send"
  | "message"
  | "minimize"
  | "note"
  | "person"
  | "persons"
  | "phone"
  | "remove"
  | "reports"
  | "search"
  | "settings"
  | "setup"
  | "split-1-right"
  | "split-2-right"
  | "tags"
  | "upload"
  | "warning"
  | "web";